Improve the Included GUI interface for switching user agents.
Use QXMLStreamReader to parse the xml file so as not to include qt xml
as a dependency and to be faster.
Delay loading the menu until first requested.
Don't query QSettings every time for the user agent as this can be
hundreds of times per web page, but instead introduce a static function
to get/set the user agent.
Move all the user agent files (source and resource) into the user agent
directory.
Remove the userAgentMenuAction and DefaultUserAgentAction and use the
existing QAction data property to hold the string.
Add a way for the user to input whatever user agent they want
Add a separator tag to the user agent xml file
Add a number of more user agent strings to the default set, mainly
Safari